+++ app/Http/Controllers/API/TeamController.php
@@ -6,7 +6,6 @@ use App\Helpers\ResponseFormatter;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CreateTeamRequest;
 use App\Http\Requests\UpdateTeamRequest;
-use App\Models\Company;
 use App\Models\Team;
 use Illuminate\Http\Request;
 
@@ -21,7 +20,8 @@ class TeamController extends Controller
     public function index(Request $request)
     {
         $limit = $request->input('limit', 10);
-        $teams = Team::query()->simplePaginate((int) $limit);
+        $companies = $request->user()->companies();
+        $teams = Team::query()->whereIn('company_id', $companies->pluck('id'))->simplePaginate((int) $limit);
 
         return ResponseFormatter::success($teams, 'Team successfully retrieved.');
     }
@@ -35,12 +35,17 @@ class TeamController extends Controller
     public function store(CreateTeamRequest $request)
     {
         try {
+            $company = $request->user()->companies()->find($request->company_id);
+
+            if ($company == null) {
+                throw new \Exception('You dont have access to this resource.', 403);
+            }
+
             if ($request->hasFile('icon')) {
                 $path = $request->file('icon')->store('public/icons');
                 $request->icon = $path;
             }
 
-            $company = Company::findOrFail($request->company_id);
             $team = $company->teams()->create($request->all());
 
             return ResponseFormatter::success($team, 'Team successfully created.');
@@ -57,6 +62,10 @@ class TeamController extends Controller
      */
     public function show(Team $team)
     {
+        if (! $this->isGranted($team)) {
+            return ResponseFormatter::error('You dont have access to this resource.', 403);
+        }
+
         return ResponseFormatter::success($team, 'Team successfully retrieved.');
     }
 
@@ -70,6 +79,10 @@ class TeamController extends Controller
     public function update(UpdateTeamRequest $request, Team $team)
     {
         try {
+            if (! $this->isGranted($team)) {
+                throw new \Exception('You dont have access to this resource.', 403);
+            }
+
             if ($request->hasFile('icon')) {
                 $path = $request->file('icon')->store('public/icons');
                 $request->icon = $path;
@@ -92,11 +105,20 @@ class TeamController extends Controller
     public function destroy(Team $team)
     {
         try {
+            if (! $this->isGranted($team)) {
+                throw new \Exception('You dont have access to this resource.', 403);
+            }
+
             $team->delete();
 
             return ResponseFormatter::success(message: 'Team successfully deleted.');
         } catch (\Throwable $th) {
             return ResponseFormatter::error($th->getMessage(), $th->getCode());
         }
+    }
+
+    private function isGranted(Team $team)
+    {
+        return request()->user()->companies()->where('id', $team->company_id)->exists();
     }
 }
